requirements.txt
-e .

aws-cdk.core
aws-cdk.aws-lambda
aws-cdk.aws-ec2
aws-cdk.aws-ecs
aws-cdk.aws-ecr-assets
aws-cdk.aws-ecs-patterns
aws-cdk.aws-ecr
aws-cdk.aws-iam
aws-cdk.aws-elasticloadbalancingv2
